+++ Xlicer_qt5/support/point3d.h
@@ -4,12 +4,12 @@
 class Point3D
 {
 public:
-    Point3D(double _x, double _y, double _z);
+    Point3D(float _x, float _y, float _z);
     Point3D();
 
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 };
 
 #endif // POINT3D_H
